Make sure `tr_sha1` calls pass lengths as int
This fixes inability to authenticate with password-protected session, and
a few more issues it seems.

+++ libtransmission/metainfo.c
@@ -469,7 +469,7 @@ tr_metainfoParseImpl (const tr_session  * session,
     {
       size_t len;
       char * bstr = tr_variantToStr (infoDict, TR_VARIANT_FMT_BENC, &len);
-      tr_sha1 (inf->hash, bstr, len, NULL);
+      tr_sha1 (inf->hash, bstr, (int) len, NULL);
       tr_sha1_to_hex (inf->hashString, inf->hash);
 
       if (infoDictLength != NULL)
